refactor(whichkey): extract panel renderer into its own file
